# doc/restapi/restapi.swagger2.json
{
  "swagger": "2.0",
  "info": {
    "version": "v1",
    "title": "API V1"
  },
  "basePath": "/",
  "paths": {
...
    "/api/Skills/find/{token}/{gameId}": {
      "get": {
        "tags": [ "Skills" ],
        "summary": "Find a Skill that matches <param name=\"token\" /> and <param name=\"gameId\" />.\r\n            \r\n            Example Usage: GET api/skills/find/SKILL_TOKEN/1",
        "operationId": "ApiSkillsFindByTokenByGameIdGet",
        "consumes": [],
        "produces": [],
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "description": "Token of Skill",
            "required": true,
            "type": "string"
          },
          {
            "name": "gameId",
            "in": "path",
            "description": "ID of the Game the Skill is for",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": { "204": { "description": "No Content" } },
        "deprecated": false
      }
    },
    "/api/Skills/find/{token}/global": {
      "get": {
        "tags": [ "Skills" ],
        "summary": "Find a Skill that matches <param name=\"token\" /> and <param name=\"gameId\" />.\r\n            \r\n            Example Usage: GET api/skills/find/SKILL_TOKEN/1",
        "operationId": "ApiSkillsFindByTokenGlobalGet",
        "consumes": [],
        "produces": [],
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "description": "Token of Skill",
            "required": true,
            "type": "string"
          },
          {
            "name": "gameId",
            "in": "path",
            "description": "ID of the Game the Skill is for",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": { "204": { "description": "No Content" } },
        "deprecated": false
      }
    },
    "/api/Skills/global/list": {
      "get": {
        "tags": [ "Skills" ],
        "summary": "Find a list of Skills that match <param name=\"gameId\" />.\r\n            If global is provided instead of a gameId, get all global skills, ie. skills that are not associated with a specific game.\r\n            \r\n            Example Usage: GET api/skills/game/1/list",
        "operationId": "ApiSkillsGlobalListGet",
        "consumes": [],
        "produces": [],
        "parameters": [
          {
            "name": "gameId",
            "in": "path",
            "description": "Game ID",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": { "204": { "description": "No Content" } },
        "deprecated": false
      }
    },
    "/api/Skills/game/{gameId}/list": {
      "get": {
        "tags": [ "Skills" ],
        "summary": "Find a list of Skills that match <param name=\"gameId\" />.\r\n            If global is provided instead of a gameId, get all global skills, ie. skills that are not associated with a specific game.\r\n            \r\n            Example Usage: GET api/skills/game/1/list",
        "operationId": "ApiSkillsGameByGameIdListGet",
        "consumes": [],
        "produces": [],
        "parameters": [
          {
            "name": "gameId",
            "in": "path",
            "description": "Game ID",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": { "204": { "description": "No Content" } },
        "deprecated": false
      }
    },
    "/api/Skills/game/{gameId}/evaluate": {
      "get": {
        "tags": [ "Skills" ],
        "summary": "Find the current progress for all skills for a <param name=\"gameId\" /> for <param name=\"actorId\" />.\r\n            \r\n            Example Usage: GET api/skills/game/1/evaluate/1",
        "operationId": "ApiSkillsGameByGameIdEvaluateGet",
        "consumes": [],
        "produces": [],
        "parameters": [
          {
            "name": "gameId",
            "in": "path",
            "description": "ID of Game",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "actorId",
            "in": "path",
            "description": "ID of Group/User",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": { "204": { "description": "No Content" } },
        "deprecated": false
      }
    },
    "/api/Skills/global/evaluate": {
      "get": {
        "tags": [ "Skills" ],
        "summary": "Find the current progress for all skills for a <param name=\"gameId\" /> for <param name=\"actorId\" />.\r\n            \r\n            Example Usage: GET api/skills/game/1/evaluate/1",
        "operationId": "ApiSkillsGlobalEvaluateGet",
        "consumes": [],
        "produces": [],
        "parameters": [
          {
            "name": "gameId",
            "in": "path",
            "description": "ID of Game",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "actorId",
            "in": "path",
            "description": "ID of Group/User",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": { "204": { "description": "No Content" } },
        "deprecated": false
      }
    },
    "/api/Skills/game/{gameId}/evaluate/{actorId}": {
      "get": {
        "tags": [ "Skills" ],
        "summary": "Find the current progress for all skills for a <param name=\"gameId\" /> for <param name=\"actorId\" />.\r\n            \r\n            Example Usage: GET api/skills/game/1/evaluate/1",
        "operationId": "ApiSkillsGameByGameIdEvaluateByActorIdGet",
        "consumes": [],
        "produces": [],
        "parameters": [
          {
            "name": "gameId",
            "in": "path",
            "description": "ID of Game",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "actorId",
            "in": "path",
            "description": "ID of Group/User",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": { "204": { "description": "No Content" } },
        "deprecated": false
      }
    },
    "/api/Skills/global/evaluate/{actorId}": {
      "get": {
        "tags": [ "Skills" ],
        "summary": "Find the current progress for all skills for a <param name=\"gameId\" /> for <param name=\"actorId\" />.\r\n            \r\n            Example Usage: GET api/skills/game/1/evaluate/1",
        "operationId": "ApiSkillsGlobalEvaluateByActorIdGet",
        "consumes": [],
        "produces": [],
        "parameters": [
          {
            "name": "gameId",
            "in": "path",
            "description": "ID of Game",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "actorId",
            "in": "path",
            "description": "ID of Group/User",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": { "204": { "description": "No Content" } },
        "deprecated": false
      }
    },
    "/api/Skills/{token}/{gameId}/evaluate": {
      "get": {
        "tags": [ "Skills" ],
        "summary": "Find the current progress for a Skill for <param name=\"actorId\" />.\r\n            \r\n            Example Usage: GET api/skills/SKILL_TOKEN/1/evaluate/1",
        "operationId": "ApiSkillsByTokenByGameIdEvaluateGet",
        "consumes": [],
        "produces": [],
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "description": "Token of Skill",
            "required": true,
            "type": "string"
          },
          {
            "name": "gameId",
            "in": "path",
            "description": "ID of the Game the Skill is for",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "actorId",
            "in": "path",
            "description": "ID of Group/User",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": { "204": { "description": "No Content" } },
        "deprecated": false
      }
    },
    "/api/Skills/{token}/global/evaluate": {
      "get": {
        "tags": [ "Skills" ],
        "summary": "Find the current progress for a Skill for <param name=\"actorId\" />.\r\n            \r\n            Example Usage: GET api/skills/SKILL_TOKEN/1/evaluate/1",
        "operationId": "ApiSkillsByTokenGlobalEvaluateGet",
        "consumes": [],
        "produces": [],
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "description": "Token of Skill",
            "required": true,
            "type": "string"
          },
          {
            "name": "gameId",
            "in": "path",
            "description": "ID of the Game the Skill is for",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "actorId",
            "in": "path",
            "description": "ID of Group/User",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": { "204": { "description": "No Content" } },
        "deprecated": false
      }
    },
    "/api/Skills/{token}/{gameId}/evaluate/{actorId}": {
      "get": {
        "tags": [ "Skills" ],
        "summary": "Find the current progress for a Skill for <param name=\"actorId\" />.\r\n            \r\n            Example Usage: GET api/skills/SKILL_TOKEN/1/evaluate/1",
        "operationId": "ApiSkillsByTokenByGameIdEvaluateByActorIdGet",
        "consumes": [],
        "produces": [],
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "description": "Token of Skill",
            "required": true,
            "type": "string"
          },
          {
            "name": "gameId",
            "in": "path",
            "description": "ID of the Game the Skill is for",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "actorId",
            "in": "path",
            "description": "ID of Group/User",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": { "204": { "description": "No Content" } },
        "deprecated": false
      }
    },
    "/api/Skills/{token}/global/evaluate/{actorId}": {
      "get": {
        "tags": [ "Skills" ],
        "summary": "Find the current progress for a Skill for <param name=\"actorId\" />.\r\n            \r\n            Example Usage: GET api/skills/SKILL_TOKEN/1/evaluate/1",
        "operationId": "ApiSkillsByTokenGlobalEvaluateByActorIdGet",
        "consumes": [],
        "produces": [],
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "description": "Token of Skill",
            "required": true,
            "type": "string"
          },
          {
            "name": "gameId",
            "in": "path",
            "description": "ID of the Game the Skill is for",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "actorId",
            "in": "path",
            "description": "ID of Group/User",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": { "204": { "description": "No Content" } },
        "deprecated": false
      }
    },
    "/api/Skills/create": {
      "post": {
        "tags": [ "Skills" ],
        "summary": "Create a new Skill.\r\n            Requires {PlayGen.SUGAR.Contracts.EvaluationCreateRequest.Name} to be unique to that {PlayGen.SUGAR.Contracts.EvaluationCreateRequest.GameId}.\r\n            \r\n            Example Usage: POST api/skills/create",
        "operationId": "ApiSkillsCreatePost",
        "consumes": [ "application/json", "text/json", "application/json-patch+json" ],
        "produces": [],
        "parameters": [
          {
            "name": "newSkill",
            "in": "body",
            "description": "{PlayGen.SUGAR.Contracts.EvaluationCreateRequest} object that holds the details of the new Skill.",
            "required": false,
            "schema": { "$ref": "#/definitions/EvaluationCreateRequest" }
          }
        ],
        "responses": { "204": { "description": "No Content" } },
        "deprecated": false
      }
    },
    "/api/Skills/update": {
      "put": {
        "tags": [ "Skills" ],
        "summary": "Update an existing Skill.\r\n            \r\n            Example Usage: PUT api/skills/update",
        "operationId": "ApiSkillsUpdatePut",
        "consumes": [ "application/json", "text/json", "application/json-patch+json" ],
        "produces": [],
        "parameters": [
          {
            "name": "skill",
            "in": "body",
            "description": "{PlayGen.SUGAR.Contracts.EvaluationCreateRequest} object that holds the details of the Skill.",
            "required": false,
            "schema": { "$ref": "#/definitions/EvaluationCreateRequest" }
          }
        ],
        "responses": { "204": { "description": "No Content" } },
        "deprecated": false
      }
    },
    "/api/Skills/{token}/global": {
      "delete": {
        "tags": [ "Skills" ],
        "summary": "Delete Skill with the <param name=\"token\" /> and <param name=\"gameId\" /> provided.\r\n            \r\n            Example Usage: DELETE api/skills/SKILL_TOKEN/1",
        "operationId": "ApiSkillsByTokenGlobalDelete",
        "consumes": [],
        "produces": [],
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "description": "Token of Skill",
            "required": true,
            "type": "string"
          },
          {
            "name": "gameId",
            "in": "path",
            "description": "ID of the Game the Skill is for",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": { "204": { "description": "No Content" } },
        "deprecated": false
      }
    },
    "/api/Skills/{token}/{gameId}": {
      "delete": {
        "tags": [ "Skills" ],
        "summary": "Delete Skill with the <param name=\"token\" /> and <param name=\"gameId\" /> provided.\r\n            \r\n            Example Usage: DELETE api/skills/SKILL_TOKEN/1",
        "operationId": "ApiSkillsByTokenByGameIdDelete",
        "consumes": [],
        "produces": [],
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "description": "Token of Skill",
            "required": true,
            "type": "string"
          },
          {
            "name": "gameId",
            "in": "path",
            "description": "ID of the Game the Skill is for",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": { "204": { "description": "No Content" } },
        "deprecated": false
      }
    },
...
  },
  "definitions": {
...
    "EvaluationCreateRequest": {
      "description": "Encapsulates achievement/skill details.",
      "required": [ "token", "name", "actorType", "completionCriteria" ],
      "type": "object",
      "properties": {
        "token": {
          "description": "A unique identifier used in development to reference the achievement/skill.",
          "type": "string"
        },
        "gameId": {
          "format": "int32",
          "description": "The ID of the Game which this achievement/skill should belong to. Can be left null to make the achievement/skill system-wide.",
          "type": "integer"
        },
        "name": {
          "description": "The display name for the achievement/skill.",
          "maxLength": 64,
          "minLength": 0,
          "type": "string"
        },
        "description": {
          "description": "The description of the achievement/skill.",
          "maxLength": 256,
          "minLength": 0,
          "type": "string"
        },
        "actorType": {
          "description": "The type of actor which this achievement/skill is intended to be completed by.",
          "enum": [ "Undefined", "User", "Group" ],
          "type": "string"
        },
        "completionCriteria": {
          "description": "A list of criteria which will be checked in order to see if an actor has completed the achievement/skill.\r\n            Must contain at least one criteria.",
          "type": "array",
          "items": { "$ref": "#/definitions/EvaluationCriteria" }
        },
        "reward": {
          "description": "A list of rewards that will be provided to the actor upon completion of the achievement/skill criteria.\r\n            An achievement does not need to contain a reward.",
          "type": "array",
          "items": { "$ref": "#/definitions/Reward" }
        }
      }
    },
    "EvaluationCriteria": {
      "description": "Encapsulates requirements for completing an achievement or skill.",
      "required": [ "key", "dataType", "criteriaQueryType", "comparisonType", "scope", "value" ],
      "type": "object",
      "properties": {
        "key": {
          "description": "The key which will be queried against to check if criteria is met.",
          "type": "string"
        },
        "dataType": {
          "description": "GameDataType of the key which is being queried.",
          "enum": [ "String", "Long", "Float", "Boolean" ],
          "type": "string"
        },
        "criteriaQueryType": {
          "description": "Which stored GameData will be queried.",
          "enum": [ "Any", "Sum", "Latest" ],
          "type": "string"
        },
        "comparisonType": {
          "description": "How the target value and the actual value will be compared.",
          "enum": [ "Equals", "NotEqual", "Greater", "GreaterOrEqual", "Less", "LessOrEqual" ],
          "type": "string"
        },
        "scope": {
          "description": "Whether the criteria will be checked against the actor or related actors (i.e. group members, user friends).",
          "enum": [ "Actor", "RelatedActors" ],
          "type": "string"
        },
        "value": {
          "description": "The value which will compared against in order to see if the criteria has been met.",
          "type": "string"
        }
      }
    },
    "Reward": {
      "description": "Encapsulates the reward given for completing an achievement or skill.",
      "required": [ "key", "dataType", "value" ],
      "type": "object",
      "properties": {
        "key": {
          "description": "The key which will be stored in GameData.",
          "type": "string"
        },
        "dataType": {
          "description": "GameDataType of the value for this GameData.",
          "enum": [ "String", "Long", "Float", "Boolean" ],
          "type": "string"
        },
        "value": {
          "description": "The value of the GameData.",
          "type": "string"
        }
      }
    },
...
  },
  "securityDefinitions": {}
}